+++ AccidentsReports/Data/Entities/RDA.cs
@@ -9,8 +9,8 @@ namespace AccidentsReports.Data.Entities {
     public class RDA {
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.None)]
-        public int EmpId { get; set; }
-        public int RDANIC { get; set; }
+        public long EmpId { get; set; }
+        public long RDANIC { get; set; }
         [Required]
         [StringLength(50)]
         public string RDADomain { get; set; }
